fix(living-documentation): generate dashboard immediately after creation
The CW alarm event-driven invocation of the generator Lambda would be
too late since it triggers when the alarm is green.

So we move from event based trigger to scheduler every minute, which is
still covered by Lambda free tier and frees us from timing pains.

There might be still some gap between initial creation of the RootMail
alarm and the first call of the dashboard generator Lambda which would
result in exceptions in the generator Lambda. This was a deliberate
tradeoff.