tests: avoid broken cleanup due to nonexisting git references
Sometimes, in Pull Requests, a branch reference might be already deleted
on GitHub, probably due to force pushes on feature branches. Then the
cleanup CodeBuild fails with

```
reference not found for primary source and source version ...
```

To workaround this, always call the cleanup CodeBuild from the default
`main` branch. This change will result in the `buildspec-cleanup.yaml`
always being used from the `main` branch, not a feature branch.